+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/utils/IDBUtils.kt
@@ -727,6 +727,8 @@ interface IDBUtils {
         }
     }
 
+    fun getPathRelativePath(context: Context, galleryId: String): String?
+
     fun getPathModifiedDate(context: Context, pathId: String): Long? {
         val columns = arrayOf(DATE_MODIFIED)
         val sortOrder = "$DATE_MODIFIED desc"